+++ src/main/java/com/example/spinlog/statistics/controller/GenderStatisticsController.java
@@ -3,11 +3,11 @@ package com.example.spinlog.statistics.controller;
 import com.example.spinlog.article.entity.RegisterType;
 import com.example.spinlog.global.response.ApiResponseWrapper;
 import com.example.spinlog.global.response.ResponseUtils;
-import com.example.spinlog.statistics.repository.dto.GenderSatisfactionAverageDto;
+import com.example.spinlog.statistics.dto.GenderSatisfactionAverageDto;
 import com.example.spinlog.statistics.service.GenderStatisticsService;
-import com.example.spinlog.statistics.service.dto.GenderDailyAmountSumResponse;
-import com.example.spinlog.statistics.service.dto.GenderEmotionAmountAverageResponse;
-import com.example.spinlog.statistics.service.dto.GenderWordFrequencyResponse;
+import com.example.spinlog.statistics.dto.response.GenderDailyAmountSumResponse;
+import com.example.spinlog.statistics.dto.response.GenderEmotionAmountAverageResponse;
+import com.example.spinlog.statistics.dto.response.GenderWordFrequencyResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.GetMapping;
