+++ src/main/java/com/example/spinlog/statistics/controller/MBTIStatisticsController.java
@@ -3,10 +3,10 @@ package com.example.spinlog.statistics.controller;
 import com.example.spinlog.article.entity.RegisterType;
 import com.example.spinlog.global.response.ApiResponseWrapper;
 import com.example.spinlog.global.response.ResponseUtils;
-import com.example.spinlog.statistics.service.dto.MBTIDailyAmountSumResponse;
-import com.example.spinlog.statistics.service.dto.MBTIEmotionAmountAverageResponse;
-import com.example.spinlog.statistics.service.dto.MBTISatisfactionAverageResponse;
-import com.example.spinlog.statistics.service.dto.MBTIWordFrequencyResponse;
+import com.example.spinlog.statistics.dto.response.MBTIDailyAmountSumResponse;
+import com.example.spinlog.statistics.dto.response.MBTIEmotionAmountAverageResponse;
+import com.example.spinlog.statistics.dto.response.MBTISatisfactionAverageResponse;
+import com.example.spinlog.statistics.dto.response.MBTIWordFrequencyResponse;
 import com.example.spinlog.statistics.service.MBTIStatisticsService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
