+++ src/main/java/com/example/spinlog/statistics/dto/GenderSatisfactionAverageDto.java
@@ -1,4 +1,4 @@
-package com.example.spinlog.statistics.repository.dto;
+package com.example.spinlog.statistics.dto;
 
 import com.example.spinlog.user.entity.Gender;
 import lombok.Builder;
