+++ src/main/java/com/example/spinlog/statistics/dto/response/MBTIEmotionAmountAverageResponse.java
@@ -1,8 +1,8 @@
-package com.example.spinlog.statistics.service.dto;
+package com.example.spinlog.statistics.dto.response;
 
 import com.example.spinlog.article.entity.Emotion;
 import com.example.spinlog.statistics.entity.MBTIFactor;
-import com.example.spinlog.statistics.repository.dto.MBTIEmotionAmountAverageDto;
+import com.example.spinlog.statistics.dto.MBTIEmotionAmountAverageDto;
 import com.example.spinlog.user.entity.Mbti;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
