+++ src/main/java/com/example/spinlog/statistics/dto/wordanalysis/WordFrequency.java
@@ -1,4 +1,4 @@
-package com.example.spinlog.statistics.service.dto;
+package com.example.spinlog.statistics.dto.wordanalysis;
 
 
 import lombok.Builder;
