+++ src/main/java/com/example/spinlog/statistics/service/MBTIStatisticsService.java
@@ -2,12 +2,12 @@ package com.example.spinlog.statistics.service;
 
 import com.example.spinlog.article.entity.Emotion;
 import com.example.spinlog.article.entity.RegisterType;
+import com.example.spinlog.statistics.dto.response.MBTIDailyAmountSumResponse;
+import com.example.spinlog.statistics.dto.response.MBTIEmotionAmountAverageResponse;
+import com.example.spinlog.statistics.dto.response.MBTISatisfactionAverageResponse;
+import com.example.spinlog.statistics.dto.response.MBTIWordFrequencyResponse;
 import com.example.spinlog.statistics.entity.MBTIFactor;
-import com.example.spinlog.statistics.repository.dto.*;
-import com.example.spinlog.statistics.service.dto.MBTIDailyAmountSumResponse;
-import com.example.spinlog.statistics.service.dto.MBTIEmotionAmountAverageResponse;
-import com.example.spinlog.statistics.service.dto.MBTISatisfactionAverageResponse;
-import com.example.spinlog.statistics.service.dto.MBTIWordFrequencyResponse;
+import com.example.spinlog.statistics.dto.*;
 import com.example.spinlog.statistics.repository.MBTIStatisticsRepository;
 import com.example.spinlog.statistics.loginService.AuthenticatedUserService;
 import com.example.spinlog.statistics.service.wordanalysis.WordExtractionService;
