+++ src/main/java/com/example/spinlog/statistics/service/wordanalysis/MorphemeApiClient.java
@@ -1,7 +1,7 @@
 package com.example.spinlog.statistics.service.wordanalysis;
 
-import com.example.spinlog.statistics.service.dto.MorphemeApiRequestDto;
-import com.example.spinlog.statistics.service.dto.MorphemeResponse;
+import com.example.spinlog.statistics.dto.wordanalysis.MorphemeApiRequest;
+import com.example.spinlog.statistics.dto.wordanalysis.MorphemeResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -12,5 +12,5 @@ import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 @FeignClient(name = "MorphemeApiClient", url = "http://aiopen.etri.re.kr:8000/WiseNLU")
 public interface MorphemeApiClient {
     @PostMapping(consumes = APPLICATION_JSON_VALUE)
-    MorphemeResponse getMorphemes(@RequestHeader("Authorization") String authorization, @RequestBody MorphemeApiRequestDto dto);
+    MorphemeResponse getMorphemes(@RequestHeader("Authorization") String authorization, @RequestBody MorphemeApiRequest dto);
 }
