+++ src/main/java/com/example/spinlog/statistics/utils/GenderStatisticsResponseMapper.java
@@ -1,10 +1,10 @@
 package com.example.spinlog.statistics.utils;
 
-import com.example.spinlog.statistics.repository.dto.GenderDailyAmountSumDto;
-import com.example.spinlog.statistics.repository.dto.GenderEmotionAmountAverageDto;
-import com.example.spinlog.statistics.repository.dto.GenderSatisfactionAverageDto;
-import com.example.spinlog.statistics.service.dto.GenderDailyAmountSumResponse;
-import com.example.spinlog.statistics.service.dto.GenderEmotionAmountAverageResponse;
+import com.example.spinlog.statistics.dto.GenderDailyAmountSumDto;
+import com.example.spinlog.statistics.dto.GenderEmotionAmountAverageDto;
+import com.example.spinlog.statistics.dto.GenderSatisfactionAverageDto;
+import com.example.spinlog.statistics.dto.response.GenderDailyAmountSumResponse;
+import com.example.spinlog.statistics.dto.response.GenderEmotionAmountAverageResponse;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
