+++ src/main/java/com/example/spinlog/statistics/utils/StatisticsZeroPaddingUtils.java
@@ -1,8 +1,8 @@
 package com.example.spinlog.statistics.utils;
 
 import com.example.spinlog.article.entity.Emotion;
-import com.example.spinlog.statistics.repository.dto.GenderDailyAmountSumDto;
-import com.example.spinlog.statistics.repository.dto.GenderEmotionAmountAverageDto;
+import com.example.spinlog.statistics.dto.GenderDailyAmountSumDto;
+import com.example.spinlog.statistics.dto.GenderEmotionAmountAverageDto;
 import com.example.spinlog.user.entity.Gender;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
