+++ src/main/resources/application.properties
@@ -2,7 +2,7 @@ spring.application.name=spinlog
 spring.profiles.active=local
 spring.data.web.pageable.one-indexed-parameters=true
 
-mybatis.type-aliases-package=com.example.spinlog.statistics.repository.dto
+mybatis.type-aliases-package=com.example.spinlog.statistics.dto
 mybatis.configuration.map-underscore-to-camel-case=true
 
 openai.url= https://api.openai.com
