+++ src/test/java/com/example/spinlog/statistics/repository/MBTIStatisticsRepositoryTest.java
@@ -5,10 +5,10 @@ import com.example.spinlog.article.entity.Emotion;
 import com.example.spinlog.article.entity.RegisterType;
 import com.example.spinlog.article.repository.ArticleRepository;
 import com.example.spinlog.statistics.entity.MBTIFactor;
-import com.example.spinlog.statistics.repository.dto.MBTIDailyAmountSumDto;
-import com.example.spinlog.statistics.repository.dto.MBTIEmotionAmountAverageDto;
-import com.example.spinlog.statistics.repository.dto.MBTISatisfactionAverageDto;
-import com.example.spinlog.statistics.repository.dto.MemoDto;
+import com.example.spinlog.statistics.dto.MBTIDailyAmountSumDto;
+import com.example.spinlog.statistics.dto.MBTIEmotionAmountAverageDto;
+import com.example.spinlog.statistics.dto.MBTISatisfactionAverageDto;
+import com.example.spinlog.statistics.dto.MemoDto;
 import com.example.spinlog.user.entity.Gender;
 import com.example.spinlog.user.entity.Mbti;
 import com.example.spinlog.user.entity.User;
