+++ src/test/java/com/example/spinlog/statistics/service/GenderStatisticsServiceTest.java
@@ -1,10 +1,13 @@
 package com.example.spinlog.statistics.service;
 
 import com.example.spinlog.article.entity.Emotion;
+import com.example.spinlog.statistics.dto.*;
+import com.example.spinlog.statistics.dto.response.GenderDailyAmountSumResponse;
+import com.example.spinlog.statistics.dto.response.GenderEmotionAmountAverageResponse;
+import com.example.spinlog.statistics.dto.response.GenderWordFrequencyResponse;
+import com.example.spinlog.statistics.dto.wordanalysis.WordFrequency;
 import com.example.spinlog.statistics.repository.GenderStatisticsRepository;
-import com.example.spinlog.statistics.repository.dto.*;
 import com.example.spinlog.statistics.service.cache.GenderStatisticsCacheFallbackService;
-import com.example.spinlog.statistics.service.dto.*;
 import com.example.spinlog.statistics.service.wordanalysis.WordExtractionService;
 import com.example.spinlog.user.entity.Gender;
 import org.assertj.core.groups.Tuple;
