+++ src/test/java/com/example/spinlog/statistics/service/cache/scheduled/GenderStatisticsCacheRefreshScheduledServiceTest.java
@@ -2,11 +2,10 @@ package com.example.spinlog.statistics.service.cache.scheduled;
 
 import com.example.spinlog.article.entity.Emotion;
 import com.example.spinlog.statistics.repository.GenderStatisticsRepository;
-import com.example.spinlog.statistics.repository.dto.GenderEmotionAmountAverageDto;
+import com.example.spinlog.statistics.dto.GenderEmotionAmountAverageDto;
 import com.example.spinlog.statistics.service.StatisticsPeriodManager;
 import com.example.spinlog.statistics.service.cache.GenderStatisticsCacheWriteService;
 import com.example.spinlog.statistics.service.fetch.GenderStatisticsRepositoryFetchService;
-import com.example.spinlog.statistics.service.cache.scheduled.GenderStatisticsCacheRefreshScheduledService;
 import com.example.spinlog.user.entity.Gender;
 import com.example.spinlog.util.CacheConfiguration;
 import com.example.spinlog.util.MockCacheHashRepository;
